+++ sales-tax-rates/src/main/java/com/complyt/domain/RatesMetaData.java
@@ -3,5 +3,5 @@ package com.complyt.domain;
 import lombok.With;
 
 @With
-public record RatesMetaData(float cityDistrictRate, float countyDistrictRate) {
+public record RatesMetaData(double cityDistrictRate, double countyDistrictRate) {
 }
