+++ sales-tax-rates/src/main/java/com/complyt/v1/model/RatesMetaDataDto.java
@@ -3,5 +3,5 @@ package com.complyt.v1.model;
 import io.swagger.v3.oas.annotations.media.Schema;
 
 @Schema(name = "RatesMetaData")
-public record RatesMetaDataDto(float cityDistrictRate, float countyDistrictRate) {
+public record RatesMetaDataDto(double cityDistrictRate, double countyDistrictRate) {
 }
