+++ sales-tax/src/main/java/com/complyt/domain/sales_tax/RatesMetaData.java
@@ -1,4 +1,4 @@
 package com.complyt.domain.sales_tax;
 
-public record RatesMetaData(float cityDistrictRate, float countyDistrictRate) {
+public record RatesMetaData(double cityDistrictRate, double countyDistrictRate) {
 }
