+++ sales-tax/src/main/java/com/complyt/domain/sales_tax/SalesTax.java
@@ -3,5 +3,5 @@ package com.complyt.domain.sales_tax;
 import lombok.*;
 
 @With
-public record SalesTax(float amount, SalesTaxRates salesTaxRates) {
+public record SalesTax(double amount, SalesTaxRates salesTaxRates) {
 }
